moodify/backend/requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.6.0
Flask-Cors==4.0.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
scikit-learn==1.4.1
pandas==2.2.1
numpy==1.26.4
requests==2.31.0
gunicorn==21.2.0
python-dateutil==2.8.2
SQLAlchemy==2.0.27
alembic==1.13.1 